make hash_xxh3_hash128 return u128
Aligns signature with hash_xxh3_hash64

+++ turbopack/crates/turbopack-ecmascript/src/chunk/content_entry.rs
@@ -21,7 +21,7 @@ use crate::chunk::{
 #[derive(Debug)]
 pub struct EcmascriptChunkContentEntry {
     pub code: ResolvedVc<Code>,
-    pub hash: ResolvedVc<u64>,
+    pub hash: ResolvedVc<u128>,
 }
 
 impl EcmascriptChunkContentEntry {
